requirements.txt
# Core language & ML
torch==2.0.1
transformers==4.36.2

# NLP
spacy==3.7.2
vaderSentiment==3.3.2
textstat==0.7.3
nltk==3.8.1

# Data handling
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.7.3
plotly==5.18.0
seaborn==0.12.2

# Backend API
fastapi==0.110.0
uvicorn==0.27.1

# Utilities
tqdm==4.66.1
pyyaml==6.0.1